Several pitch related improvemnts
* Rewritten resampler class that does a much better jobb at converting audio
from 32kHz to other sample rates
* Moved from cubic interpolation to linear interpolation in the WaveOscillator
class. This makes no noticable degradation in audio quality and is correct
according to how it is implemented on the read hardware. It also reduces the
CPU usage.
* Fixes a bug that created a lot of ticking noise on certain instruments
* Some other minor fixes

+++ libemusc/src/pitch.h
@@ -46,7 +46,7 @@ public:
 
   void note_off();
 
-  inline int get_phase_increment(void) {
+  inline float get_phase_increment(void) {
     _currentInc += _deltaInc; return _currentInc; }
 
   inline uint16_t get_sample_id(void) { return _sampleIndex; }
